+++ src/collections/index.ts
@@ -2,6 +2,7 @@ import { Branches } from "./branches.js";
 import { Comments } from "./comments.js";
 import { Contributors } from "./contributors.js";
 import { Files } from "./files.js";
+import { GlossaryTerms } from "./glossary_terms.js";
 import { Jwt } from "./jwt.js";
 import { Keys } from "./keys.js";
 import { Languages } from "./languages.js";
@@ -28,6 +29,7 @@ export {
 	Comments,
 	Contributors,
 	Files,
+	GlossaryTerms,
 	Jwt,
 	Keys,
 	Languages,